pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-note-triage"
version = "0.1.0"
description = "A keyboard-first PySide6 app for triaging Markdown notes with DeepSeek-powered suggestions."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Zoet2004" }
]
keywords = ["markdown", "notes", "pyside6", "llm", "siliconflow", "deepseek"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Win32 (MS Windows)",
  "Environment :: X11 Applications :: Qt",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Desktop Environment",
  "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
  "markdown",
  "openai",
  "PySide6",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "build",
  "pytest",
  "pytest-cov",
  "ruff",
]
package = [
  "pyinstaller",
]

[project.scripts]
md-triage = "src.app:main"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
